+++ src/main/java/com/example/carpj/entity/Car.java
@@ -19,7 +19,7 @@ public class Car {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cno;        // 번호 1,2,3~~~9999
+    private Long cno;        // 번호 1,2,3~~~9999
     private String platform;   // 중고차 플랫폼 Ex. kb차차차, kcar
     private String tag;         // 국산 / 수입 태그
     private String type;        // 차종 Ex. suv, 경차, 대형차 등
